==> 10_data.qmd <==
# Data

Data used for TROLL evaluation are composed of data used as inputs for TROLL simulations and data used to assess the simulation outputs. TROLL runs with species parameters, including functional traits and allometric parameters, climate data, and soil data. We evaluated TROLL simulations against forest fluxes and final forest structure and composition. Data include thus forest inventories (species inputs, forest structure and composition evaluation), functional traits (species inputs and forest composition evaluation), soil parameters (soil inputs), climate data (climate input), and fluxes data (forest fluxes evaluation).
